Refactor logic to contain in one stimulus controller, add anchor page detecting behavior
