+++ app/components/sidebar/link_component.html.erb
@@ -1,4 +1,4 @@
-<li class="<%= @class %>" data-sidebar-group-target="link" data-sidebar-anchor-menu-target="link">
+<li class="<%= @class %>" data-sidebar-group-target="link">
   <%= link_to @path do %>
     <% if @icon %>
       <i class="lni mr-10 lni-<%= @icon %>"></i>
