+++ app/javascript/controllers/index.js
@@ -8,11 +8,11 @@ import AlertController from './alert_controller'
 
 import AutosaveController from './autosave_controller'
 
-import CourtOrderFormController from './court_order_form_controller'
-
 import DisableFormController from './disable_form_controller'
 
 import DismissController from './dismiss_controller'
+
+import CourtOrderFormController from './court_order_form_controller'
 
 import HelloController from './hello_controller'
 
@@ -24,8 +24,6 @@ import NavbarController from './navbar_controller'
 
 import SelectAllController from './select_all_controller'
 
-import SidebarAnchorLinkController from './sidebar_anchor_menu_controller'
-
 import SidebarController from './sidebar_controller'
 
 import SidebarGroupController from './sidebar_group_controller'
@@ -33,15 +31,14 @@ import SidebarGroupController from './sidebar_group_controller'
 import TruncatedTextController from './truncated_text_controller'
 application.register('alert', AlertController)
 application.register('autosave', AutosaveController)
-application.register('court-order-form', CourtOrderFormController)
 application.register('disable-form', DisableFormController)
 application.register('dismiss', DismissController)
+application.register('court-order-form', CourtOrderFormController)
 application.register('hello', HelloController)
 application.register('icon-toggle', IconToggleController)
 application.register('multiple-select', MultipleSelectController)
 application.register('navbar', NavbarController)
 application.register('select-all', SelectAllController)
-application.register('sidebar-anchor-menu', SidebarAnchorLinkController)
 application.register('sidebar', SidebarController)
 application.register('sidebar-group', SidebarGroupController)
 application.register('truncated-text', TruncatedTextController)
