+++ app/javascript/controllers/sidebar_group_controller.js
@@ -4,16 +4,68 @@ export default class extends Controller {
   static targets = ['title', 'list', 'link']
 
   connect () {
-    this.toggleShow()
+    const performToggleShow = this.linkTargets.find((link) => {
+      return link.classList.contains('active')
+    }) || this.isAnchorGroupPage()
+
+    if (performToggleShow) {
+      this.toggleShow()
+    }
+
+    if (this.isAnchorGroupPage()) {
+      this.initializeMenuHighlight()
+    }
   }
 
   // Expands list if a link is active
   toggleShow () {
-    this.linkTargets.forEach((link) => {
-      if (link.classList.contains('active')) {
-        this.titleTarget.classList.remove('collapsed')
-        this.listTarget.classList.add('show')
-      }
-    })
+    this.titleTarget.classList.remove('collapsed')
+    this.listTarget.classList.add('show')
   }
+
+    initializeMenuHighlight () {
+      const linkTargetsMap = () => {
+        const hash = {}
+        this.linkTargets.forEach(link => {
+          const href = link.firstElementChild.href
+          if (href.includes("#")) {
+            const headerId = href.substring(href.indexOf('#') + 1)
+            hash[headerId] = link
+          }
+        })
+        return hash
+      }
+  
+      const linkHash = linkTargetsMap()
+  
+      const observer = new IntersectionObserver(entries => {
+        entries.forEach(entry => {
+          if (entry.intersectionRatio > 0) {
+            this.linkTargets.forEach(link => link.classList.remove('active'))
+  
+            const headerId = entry.target.id
+            const activeLink = linkHash[headerId]
+            if (activeLink) {
+              activeLink.classList.add('active')
+            }
+          }
+        })
+      })
+  
+      document.querySelectorAll('h1[id], h2[id]').forEach((header) => {
+        observer.observe(header)
+      })
+    }
+
+    isAnchorGroupPage () {
+      if (this.linkTarget.firstElementChild) {
+        const href = this.linkTarget.firstElementChild.href
+        if (href.includes('#')) {
+          const hrefAnchorString =  href.substring(href.indexOf('#') + 1)
+          if (document.getElementById(hrefAnchorString)) {
+            return true
+          }
+        }
+      }
+    }
 }
